+++ packages/reactabular-virtualized/CHANGELOG.md
@@ -1,4 +1,9 @@
-2.0.5 / 2016-08-24
+3.0.1 / 2016-09-01
+==================
+
+  * Bug fix - Keep header and body in sync when scrolling at header.
+
+3.0.0 / 2016-09-01
 ==================
 
   * Initial release.